WC Enable buggy proc-as-args
Passing procs as args is now possible as they are treated
(somewhat) like variables. The new issue, though, is that the
way that arguments are added (when a method is called instead of
when it is defined) is really just a lousy way to do things.
Thus, going forward all arguments to function need to be specified
at its definition.